+++ R/utilities-individual.R
@@ -83,7 +83,7 @@ createDistributions <- function(individualCharacteristics) {
   )
 }
 
-#' Creates an individual using the PK-Sim Database.
+#' Create the characteristics of an individual using the PK-Sim Database.
 #'
 #' @param species Species of the individual as defined in PK-Sim (see Species enum)
 #' @param population Population to use to create the individual. This is required only when the species is Human. (See HumanPopulation enum)
@@ -97,14 +97,14 @@ createDistributions <- function(individualCharacteristics) {
 #' @param gestationalAge Gestational age of the created individual (for human species only using the Preterm population). Default is 40 Weeks
 #' @param gestationalAgeUnit Unit in which the gestational age value is defined. Default is week(s)
 #' @param moleculeOntogenies Optional list of `MoleculeOntogeny` that will be used to retrieve ontogeny information for molecules.
-#' A `MoleculeOntogeny` is an object with the name a `molecule` property (e.g the name of the molecule as defined in your simulation)
+#' A `MoleculeOntogeny` is an object with a `molecule` property (e.g. the name of the molecule as defined in your simulation)
 #' and an `ontogeny` property (e.g. the name of the predefined ontogeny to use for this molecule). The list of all available ontogenies
 #' can be accessed programmatically using the enum `StandardOntogeny`
 #' @param seed Optional seed parameter to use to generate start values for the created individual algorithm.
 #'
 #' @import ospsuite.utils
 #'
-#' @return An array of `ParameterValue` containing the value of each individual parameter
+#' @return An `IndividualCharacteristics` object.
 #'
 #' @export
 createIndividualCharacteristics <- function(