+++ tests/testthat/test-1-plot-with-ospsuite-plots-population-time-profile.R
@@ -1,13 +1,3 @@
-# Set defaults
-ggplot2::theme_update(legend.title = ggplot2::element_blank())
-ggplot2::theme_update(legend.position = c(0.95, 0.95))
-ggplot2::theme_update(legend.justification = c("right", "top"))
-
-ospsuite.plots::setOspsuite.plots.option(
-  ospsuite.plots::OptionKeys$defaultPercentiles,
-  c(0.05, 0.5, 0.95)
-)
-
 # Load simulation
 sim <- loadSimulation(
   aciclovirSimulationPath,